third_party/blink/web_tests/wpt_internal/display-lock/rendersubtree/containment/nested-containment.html

<!doctype HTML>
<html>
<meta charset="utf8">
<title>Display Locking: containment checks with nested locks</title>
<link rel="author" title="Rakina Zata Amni" href="mailto:rakina@chromium.org">
<link rel="help" href="https://github.com/WICG/display-locking">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<body>
<style>
.lockable, .parentOfLockable > div {
  contain: style layout;
}
.randomClass > div {
  color: red;
}
</style>
<script>
function forceLayout(el) {
  el.offsetTop;
}

function createParentAndChild() {
  let parent = document.createElement("div");
  let child = document.createElement("div");
  parent.appendChild(child);
  document.body.appendChild(parent);
  parent.classList.add("lockable");
  forceLayout(child);
  return parent;
}

async_test(async(t) => {
  let parent = createParentAndChild();
  let child = parent.firstChild;
  await parent.displayLock.acquire({timeout: Infinity});
  t.step(() => assert_equals(parent.displayLock.locked, true));

  let childPromise = child.displayLock.acquire({timeout: Infinity});
  childPromise.then(() => {
    t.step(() => assert_equals(child.displayLock.locked, true));
    t.done();
  });
}, "Nested locked element with clean style should not enforce containment requirement");

async_test(async(t) => {
  let parent = createParentAndChild();
  let child = parent.firstChild;
  await parent.displayLock.acquire({timeout: Infinity});
  t.step(() => assert_equals(parent.displayLock.locked, true));

  // Will mark #child for style recalc.
  parent.classList.add("randomClass");
  let childPromise = child.displayLock.acquire({timeout: Infinity});
  childPromise.then(() => {
    // Lifecycle update has finished but since #child is within a locked subtree,
    // its style is still dirty and we keep it locked.
    t.step(() => assert_equals(child.displayLock.locked, true));
    forceLayout(child);
    // Gets unlocked after forced recalc on #child.
    t.step(() => assert_equals(child.displayLock.locked, false));
    t.done();
  });
}, "Nested locked element that needs recalc should not enforce containment requirement");

async_test(async(t) => {
  let parent = createParentAndChild();
  let child = parent.firstChild;
  await parent.displayLock.acquire({timeout: Infinity});
  t.step(() => assert_equals(parent.displayLock.locked, true));

  // Will mark #child for style recalc.
  parent.classList.add("parentOfLockable");
  let childPromise = child.displayLock.acquire({timeout: Infinity});
  childPromise.then(() => {
    t.step(() => assert_equals(child.displayLock.locked, true));
    forceLayout(child);
    // Is still locked after forced recalc on #child because of added containment.
    t.step(() => assert_equals(child.displayLock.locked, true));
    t.done();
  });
}, "Nested locked element recalcs correctly when forced");
</script>
</body>
